% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massbankParser.R
\name{.massbankParseRecord}
\alias{.massbankParseRecord}
\title{MassBank parser - line parser}
\usage{
.massbankParseRecord(lines)
}
\arguments{
\item{lines}{Character vector with the record in line-by-line format.}
}
\value{
List of root-level tag entries (name: tag name, content: unprocessed tag value)
}
\description{
MassBank parser - line parser
}
